Finalized

+++ db/migrate/20150603195403_create_invoices.rb
@@ -1,7 +1,7 @@
 class CreateInvoices < ActiveRecord::Migration
   def change
     create_table :invoices do |t|
-      t.float 	:purchase_id
+      t.integer 	:purchase_id
       t.float	:debt
 
       t.timestamps null: false
